Move the append related tests to a dedicated module
